Use ruff for linting, add type annotations, break `main()` into smaller bits
ruff is really neat and super quick at checking Python code, and can check for
a whole load of various things.

Type annotations are here to stay. We should make use of them so the code is
easier to follow.

`main()` was too long, so parts of the logic from it was broken into smaller
utility functions.

The workaround which established a new Kubernetes client to work around
shortlived API tokens has been removed as the Kubernetes client handles this
correctly now.