Refactoring to make the code easier to work with
This splits the code into multiple files and smaller functions which each does
a small piece of work.

As part of this, a few new things were added/changed:

* It's now possible to override the random suffix for the created pod so it is
clearer what the pod is used for. This is especially nice if copypod is used a
lot and you end up with a bunch of pods with non-descriptive names.

* Environment variables can be set for the new pod.

* The "creator" and "original-pod" labels have been moved to annotations.

* A new "copypod" label is added with a value of "true" to make it easy to
filter copied pods.

* The "karpenter.sh/do-not-disrupt" annotation is set with a value of "true" to
make it less likely for the pod to be terminated untimely in clusters using
Karpenter for auto scaling.